Fixed tests with comments
